Added recursive flag
